Move generate Testcases outside of box
