+++ generateTestcases/generateIndices.js
