+++ generateTestcases/generatePossibleOperations/generateDeleteBlockOperations.js
