+++ generateTestcases/generatePossibleOperations/generateDeleteContentInBlockOperations.js
