+++ generateTestcases/generatePossibleOperations/generateMergeBlockOperations.js
