+++ generateTestcases/generatePossibleOperations/generateMoveBlockOperations.js
