+++ generateTestcases/generatePossibleOperations/generateSplitBlockOperations.js
