+++ generateTestcases/operations.js
