+++ generateTestcases/operations/deleteBlock.js
