+++ generateTestcases/operations/deleteContentInBlock.js
