+++ generateTestcases/operations/mergeBlocks.js
