+++ generateTestcases/operations/moveBlock.js
